client-cli/requirements.txt
requests
pyyaml
python-dotenv
websockets